flask/requirement.txt
Flask == 2.3.2
Flask-Migrate == 4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.6
SQLAlchemy==2.0.17
python-dotenv==1.0.0
psycopg2==2.9.6